% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trophic_ecology_mesopelagic-package.R
\docType{package}
\name{trophic_ecology_mesopelagic-package}
\alias{trophic_ecology_mesopelagic}
\alias{trophic_ecology_mesopelagic-package}
\title{trophic_ecology_mesopelagic: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lizloutrage/trophic_ecology_mesopelagic}
  \item Report bugs at \url{https://github.com/lizloutrage/trophic_ecology_mesopelagic/issues}
}

}
\author{
\strong{Maintainer}: Liz Loutrage \email{liz.loutrage@gmail.com} (\href{https://orcid.org/{{orcid}}}{ORCID}) [copyright holder]

}
\keyword{internal}
